refactor: split main.go into scan/warroom/pages/server files
Move-only refactor, no logic changes. main.go reduced from ~87KB/3000
lines to 144 lines (flags, store init, entry point only).

- scan.go: dashboardState, scan loop, periodic refresh
- warroom.go: War Room collection, ranking, rendering
- pages.go: infrastructure/namespaces/cost/security/waste/optimizations pages
- server.go: server struct, route registration, shared helpers, embed directive
